Extract observation handling
