+++ backend/src/main/java/ai/verta/modeldb/experimentRun/FutureExperimentRunDAO.java
@@ -3,20 +3,15 @@ package ai.verta.modeldb.experimentRun;
 import ai.verta.common.KeyValue;
 import ai.verta.common.ModelDBResourceEnum;
 import ai.verta.modeldb.*;
-import ai.verta.modeldb.common.CommonUtils;
 import ai.verta.modeldb.common.connections.UAC;
-import ai.verta.modeldb.common.exceptions.ModelDBException;
 import ai.verta.modeldb.common.exceptions.NotFoundException;
 import ai.verta.modeldb.common.futures.FutureGrpc;
 import ai.verta.modeldb.common.futures.FutureJdbi;
 import ai.verta.modeldb.common.futures.InternalFuture;
-import ai.verta.modeldb.exceptions.InvalidArgumentException;
 import ai.verta.modeldb.exceptions.PermissionDeniedException;
 import ai.verta.modeldb.experimentRun.subtypes.KeyValueHandler;
-import ai.verta.modeldb.utils.ModelDBUtils;
+import ai.verta.modeldb.experimentRun.subtypes.ObservationHandler;
 import ai.verta.uac.*;
-import com.google.protobuf.InvalidProtocolBufferException;
-import com.google.protobuf.Value;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -33,6 +28,7 @@ public class FutureExperimentRunDAO {
 
   private final KeyValueHandler metricsHandler;
   private final KeyValueHandler hyperparametersHandler;
+  private final ObservationHandler observationHandler;
 
   public FutureExperimentRunDAO(Executor executor, FutureJdbi jdbi, UAC uac) {
     this.executor = executor;
@@ -41,6 +37,7 @@ public class FutureExperimentRunDAO {
 
     metricsHandler = new KeyValueHandler(executor, jdbi, "metrics");
     hyperparametersHandler = new KeyValueHandler(executor, jdbi, "hyperparameters");
+    observationHandler = new ObservationHandler(executor, jdbi);
   }
 
   public InternalFuture<List<Observation>> getObservations(GetObservations request) {
@@ -52,53 +49,8 @@ public class FutureExperimentRunDAO {
     // Check permissions
     var currentFuture = checkPermission(runId, ModelDBActionEnum.ModelDBServiceActions.READ);
 
-    // Validate input
-    currentFuture =
-        currentFuture.thenRun(
-            () -> {
-              if (key.isEmpty()) {
-                throw new InvalidArgumentException("Empty observation key");
-              }
-            },
-            executor);
-
-    // Query
     return currentFuture.thenCompose(
-        unused ->
-            jdbi.withHandle(
-                handle ->
-                    handle
-                        .createQuery(
-                            "select k.kv_value value, k.value_type type, o.epoch_number epoch from "
-                                + "(select keyvaluemapping_id, epoch_number from observation "
-                                + "where experiment_run_id =:run_id and entity_name = \"ExperimentRunEntity\") o, "
-                                + "(select id, kv_value, value_type from keyvalue where kv_key =:name and entity_name IS NULL) k "
-                                + "where o.keyvaluemapping_id = k.id")
-                        .bind("run_id", runId)
-                        .bind("name", key)
-                        .map(
-                            (rs, ctx) -> {
-                              try {
-                                return Observation.newBuilder()
-                                    .setEpochNumber(
-                                        Value.newBuilder().setNumberValue(rs.getLong("epoch")))
-                                    .setAttribute(
-                                        KeyValue.newBuilder()
-                                            .setKey(key)
-                                            .setValue(
-                                                (Value.Builder)
-                                                    CommonUtils.getProtoObjectFromString(
-                                                        rs.getString("value"), Value.newBuilder()))
-                                            .setValueTypeValue(rs.getInt("type")))
-                                    .build();
-                              } catch (InvalidProtocolBufferException e) {
-                                LOGGER.error(
-                                    "Error generating builder for {}", rs.getString("value"));
-                                throw new ModelDBException(e);
-                              }
-                            })
-                        .list()),
-        executor);
+        unused -> observationHandler.getObservations(runId, key), executor);
   }
 
   public InternalFuture<Void> logObservations(LogObservations request) {
@@ -111,100 +63,9 @@ public class FutureExperimentRunDAO {
     // Check permissions
     var currentFuture = checkPermission(runId, ModelDBActionEnum.ModelDBServiceActions.UPDATE);
 
-    // Validate input
     currentFuture =
-        currentFuture.thenRun(
-            () -> {
-              for (final var observation : observations) {
-                if (observation.getAttribute().getKey().isEmpty()) {
-                  throw new InvalidArgumentException("Empty observation key");
-                }
-              }
-            },
-            executor);
-
-    // Log observations
-    for (final var observation : observations) {
-      final var attribute = observation.getAttribute();
-      currentFuture =
-          currentFuture
-              .thenCompose(
-                  unused -> {
-                    // If the epoch is specified, save it directly
-                    if (observation.hasEpochNumber()) {
-                      if (observation.getEpochNumber().getKindCase()
-                          != Value.KindCase.NUMBER_VALUE) {
-                        String invalidEpochMessage =
-                            "Observations can only have numeric epoch_number, condition not met in "
-                                + observation;
-                        throw new InvalidArgumentException(invalidEpochMessage);
-                      }
-                      return InternalFuture.completedInternalFuture(
-                          (long) observation.getEpochNumber().getNumberValue());
-                    } else {
-                      // Otherwise, infer at runtime. We can't do this in the same SQL command as
-                      // we'll be updating these tables and some SQL implementations don't support
-                      // select together with updates
-                      final var sql =
-                          "select max(o.epoch_number) from "
-                              + "(select keyvaluemapping_id, epoch_number from observation "
-                              + "where experiment_run_id =:run_id and entity_name = \"ExperimentRunEntity\") o, "
-                              + "(select id from keyvalue where kv_key =:name and entity_name IS NULL) k "
-                              + "where o.keyvaluemapping_id = k.id";
-                      return jdbi.withHandle(
-                          handle ->
-                              handle
-                                  .createQuery(sql)
-                                  .bind("run_id", runId)
-                                  .bind("name", attribute.getKey())
-                                  .mapTo(Long.class)
-                                  .findOne()
-                                  .map(x -> x + 1)
-                                  .orElse(0L));
-                    }
-                  },
-                  executor)
-              .thenCompose(
-                  epoch ->
-                      // Insert into KV table
-                      jdbi.useHandle(
-                          handle -> {
-                            final var kvId =
-                                handle
-                                    .createUpdate(
-                                        "insert into keyvalue (field_type, kv_key, kv_value, value_type) "
-                                            + "values (\"attributes\", :key, :value, :type)")
-                                    .bind("key", attribute.getKey())
-                                    .bind(
-                                        "value",
-                                        ModelDBUtils.getStringFromProtoObject(attribute.getValue()))
-                                    .bind("type", attribute.getValueTypeValue())
-                                    .executeAndReturnGeneratedKeys()
-                                    .mapTo(Long.class)
-                                    .one();
-
-                            // Insert to observation table
-                            // We don't need transaction here since it's fine to add to the kv table
-                            // and fail to insert into the observation table, as the value will be
-                            // just ignored
-                            handle
-                                .createUpdate(
-                                    "insert into observation (entity_name, field_type, timestamp, experiment_run_id, keyvaluemapping_id, epoch_number) "
-                                        + "values (\"ExperimentRunEntity\", \"observations\", :timestamp, :run_id, :kvid, :epoch)")
-                                .bind(
-                                    "timestamp",
-                                    observation.getTimestamp() == 0
-                                        ? now
-                                        : observation.getTimestamp())
-                                .bind("run_id", runId)
-                                .bind("kvid", kvId)
-                                .bind("epoch", epoch)
-                                .executeAndReturnGeneratedKeys()
-                                .mapTo(Long.class)
-                                .one();
-                          }),
-                  executor);
-    }
+        currentFuture.thenCompose(
+            unused -> observationHandler.logObservations(runId, observations, now), executor);
 
     return currentFuture.thenCompose(unused -> updateModifiedTimestamp(runId, now), executor);
   }
